+++ src/ViewInstanceWrapper.js
@@ -1,25 +1,25 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { omit } from 'lodash';
 
 import { checkIfUserInMemberTenant } from '@folio/stripes/core';
 import { withTags } from '@folio/stripes/smart-components';
 
 import ViewInstance from './ViewInstance';
 import { useUserTenantPermissions } from './hooks';
-import { checkIfSharedInstance } from './utils';
+import { useInstance } from './common';
 
 const ViewInstanceWrapper = (props) => {
   const {
     match: { params: { id } },
-    resources,
     stripes,
   } = props;
 
-  const instance = resources.instance?.records?.[0];
-  const selectedInstance = instance?.id === id ? instance : null;
   const userId = stripes?.user?.user?.id;
   const centralTenantId = stripes.user.user?.consortium?.centralTenantId;
+  const { instance } = useInstance(id);
+
+  const isShared = Boolean(instance?.shared);
+  const tenantId = instance?.tenantId ?? stripes.okapi.tenant;
 
   const {
     userPermissions: centralTenantPermissions,
@@ -28,34 +28,22 @@ const ViewInstanceWrapper = (props) => {
     userId,
     tenantId: centralTenantId,
   }, {
-    enabled: userId && centralTenantId && checkIfUserInMemberTenant(stripes) && checkIfSharedInstance(stripes, instance),
+    enabled: Boolean(isShared && checkIfUserInMemberTenant(stripes)),
   });
 
   return (
     <ViewInstance
-      {...omit(props, ['resources', 'mutator'])}
-      selectedInstance={selectedInstance}
+      {...props}
+      isShared={isShared}
+      tenantId={tenantId}
+      selectedInstance={instance}
       centralTenantPermissions={centralTenantPermissions}
       isCentralTenantPermissionsLoading={isCentralTenantPermissionsLoading}
     />
   );
 };
 
-ViewInstanceWrapper.manifest = Object.freeze({
-  instance: {
-    type: 'okapi',
-    path: 'inventory/instances/:{id}',
-    resourceShouldRefresh: true,
-    throwErrors: false,
-  },
-});
-
 ViewInstanceWrapper.propTypes = {
-  resources: PropTypes.shape({
-    instance: PropTypes.shape({
-      records: PropTypes.arrayOf(PropTypes.object),
-    }),
-  }).isRequired,
   stripes: PropTypes.object.isRequired,
   match: PropTypes.shape({
     params: PropTypes.shape({