+++ src/common/hooks/index.js
@@ -4,6 +4,8 @@ export { default as useGoBack } from './useGoBack';
 export { default as useFacets } from './useFacets';
 export { default as useHolding } from './useHolding';
 export { default as useInstanceQuery } from './useInstanceQuery';
+export { default as useSearchInstanceByIdQuery } from './useSearchInstanceByIdQuery';
+export { default as useTenantKy } from './useTenantKy';
 export { default as useDefaultJobProfile } from './useDefaultJobProfile';
 export { default as useAllowedJobProfiles } from './useAllowedJobProfiles';