+++ src/components/ViewSource/ViewSource.js
@@ -12,20 +12,22 @@ import {
 import { useStripes } from '@folio/stripes/core';
 import MarcView from '@folio/quick-marc/src/QuickMarcView/QuickMarcView';
 import PrintPopup from '@folio/quick-marc/src/QuickMarcView/PrintPopup';
+import { getHeaders } from '@folio/quick-marc/src/QuickMarcEditor/utils';
 
-import {
-  useInstance,
-  useGoBack,
-} from '../../common/hooks';
-import { checkIfSharedInstance, isUserInConsortiumMode } from '../../utils';
+import { useGoBack } from '../../common/hooks';
+
+import { isUserInConsortiumMode } from '../../utils';
 import MARC_TYPES from './marcTypes';
 
 import styles from './ViewSource.css';
 
 const ViewSource = ({
   mutator,
+  instance = {},
   instanceId,
+  isInstanceLoading,
   holdingsRecordId,
+  tenantId,
   marcType,
 }) => {
   const stripes = useStripes();
@@ -47,12 +49,12 @@ const ViewSource = ({
   const [marc, setMarc] = useState();
   const [isMarcLoading, setIsMarcLoading] = useState(true);
 
-  const { instance, isLoading: isInstanceLoading } = useInstance(instanceId, mutator.marcInstance);
-
   useEffect(() => {
     setIsMarcLoading(true);
 
-    mutator.marcRecord.GET()
+    const { okapi: { tenant, token, locale } } = stripes;
+
+    mutator.marcRecord.GET({ headers: getHeaders(tenantId ?? tenant, token, locale) })
       .then((marcResponse) => {
         setMarc(marcResponse);
       })
@@ -84,7 +86,7 @@ const ViewSource = ({
     <FormattedMessage
       id={`ui-inventory.marcSourceRecord.${marcType}`}
       values={{
-        shared: isUserInConsortiumMode(stripes) ? checkIfSharedInstance(stripes, instance) : null,
+        shared: isUserInConsortiumMode(stripes) ? instance.shared : null,
       }}
     />
   );
@@ -123,7 +125,10 @@ const ViewSource = ({
 ViewSource.propTypes = {
   mutator: PropTypes.object.isRequired,
   instanceId: PropTypes.string.isRequired,
+  instance: PropTypes.object,
+  isInstanceLoading: PropTypes.bool,
   holdingsRecordId: PropTypes.string,
+  tenantId: PropTypes.string,
   marcType: PropTypes.oneOf(Object.values(MARC_TYPES)).isRequired,
 };