+++ src/utils.js
@@ -21,7 +21,6 @@ import {
 } from 'lodash';
 import moment from 'moment';
 
-import { checkIfUserInCentralTenant } from '@folio/stripes/core';
 import { FormattedUTCDate } from '@folio/stripes/components';
 
 import {
@@ -782,6 +781,4 @@ export const isMARCSource = (source) => {
 
 export const isUserInConsortiumMode = stripes => stripes.hasInterface('consortia');
 
-export const checkIfSharedInstance = (stripes, instance = {}) => {
-  return instance?.source?.includes(CONSORTIUM_PREFIX) || checkIfUserInCentralTenant(stripes);
-};
+export const isInstanceShadowCopy = (source) => [`${CONSORTIUM_PREFIX}FOLIO`, `${CONSORTIUM_PREFIX}MARC`].includes(source);